#77 Refactor component structure
