+++ app/src/main/ui/src/pages/EditUser.tsx
@@ -1,9 +1,10 @@
 import {H1, LoadingIcon} from "@cloudogu/ces-theme-tailwind";
-import {useSetPageTitle} from "../../hooks/useSetPageTitle";
-import UserForm from "../users/UserForm";
-import {useUser} from "../../hooks/useUser";
-import {User} from "../../services/Users";
+import {useSetPageTitle} from "../hooks/useSetPageTitle";
+import UserForm from "../components/users/UserForm";
+import {useUser} from "../hooks/useUser";
+import {User} from "../services/Users";
 import {useParams} from "react-router-dom";
+import {t} from "../helpers/i18nHelpers";
 
 export default function EditUser(props: { title: string }) {
     useSetPageTitle(props.title);
@@ -11,7 +12,7 @@ export default function EditUser(props: { title: string }) {
     const {setUser, user, isLoading} = useUser(username);
 
     return <>
-        <H1 className="uppercase">Account</H1>
+        <H1 className="uppercase">{t("pages.usersEdit")}</H1>
         {isLoading ?
             <div className={"flex row justify-center w-[100%] mt-16"}>
                 <LoadingIcon className={"w-64 h-64"}/>
